backend/requirements.txt
fastapi==0.128.0
httpx==0.28.1
python-dotenv==1.2.1
jinja2==3.1.6
geopy==2.4.1
osmnx==2.0.7
networkx==3.6.1